refactor(admin): split analytics section modules
Why:
- keep the Quantercise analytics route composition small and stable
- separate summary panels from the daily active users chart so future changes land in focused modules

Verify:
- pnpm exec prettier --check apps/admin/src/app/quantercise/analytics/page.tsx apps/admin/src/app/quantercise/analytics/analytics-content.tsx apps/admin/src/app/quantercise/analytics/analytics-sections.tsx apps/admin/src/app/quantercise/analytics/analytics-activity-chart.tsx apps/admin/src/app/quantercise/analytics/analytics-summary-panels.tsx
- git diff --check -- apps/admin/src/app/quantercise/analytics/page.tsx apps/admin/src/app/quantercise/analytics/analytics-content.tsx apps/admin/src/app/quantercise/analytics/analytics-sections.tsx apps/admin/src/app/quantercise/analytics/analytics-activity-chart.tsx apps/admin/src/app/quantercise/analytics/analytics-summary-panels.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
